extracted ui elements into individual components
