+++ src/js/views/addbutton.js
@@ -1,0 +1,5 @@
+export const addNoteButton = `
+    <button class="add_interval_button p-3 border border-mid mx-auto rounded text-mid 2xl:text-2xl font-medium justify-self-end self-end mb-3">
+    Add Color
+    </button>
+  `;
